adjust token types to avoid C++ problems
| TIdent(s,i1)::(Tclass i2|Tstruct i2)::_,_

This case makes the name of a class or struct a typedef, so that it can be
used as a function return type in the same file.  Having ident as a
possible type makes too many conflicts.

| TIdent(s,i1)::TOPar i2::x::_,_

This case enables the parsing of constructors.  Name ( where Name is a
typedef could be the start of a function pointer declaration.  One has to
look one token further to see that there is no * after the ( to know that
is it a constructor declaration.  To avoid this problem, we convert the
typedef (identified in the previous rule) back to an identifier for this
case.

This may prevent parsing declarations of function pointers that return a
structure or class value, but perhaps function pointers are not common in
C++ anyway.