ignore bef information for meta decl
The bef information on a Decl is the place for adding something that
should come before the entire declaration, including eg static that
might not be mentioned in the semantic patch.

This feature is not needed in the MetaDecl case, because the MetaDecl
matches the whole declaaration, whatever it is.  Indeed, the bef
information is just passed along without being processed in
cocci_vs_c.ml.  So just skip it in insert_plus in finding the places
that tokens can be attached to.

Reported by Jan Tojnar in github issue #281