add // detection
When a semantic patch adds a comment //, then we have to check that this
is not immediately followed by some token from the original program,
because that token will be consumed by the command.

If there is such a token, then add a newline.  If there is already a
newline, then nothing needs to be done.