adjust treatment of patching_failed
Some changes:

* Drop the --keep-trying flag.  It could be inferred that without the
flag not all of the files are treated.  All files are always treated.
The only question is whether the files on disk are modified.

* Make the patching_failed functional.  This is necessary to pass the
result out when parmap is used.

* Improve some indentation to make it easier to see what code can see
what variables.

The improved behavior of patching_failed can be seen with the semantic
patch fails.cocci.  One can run this on some subdirectory of the Linux
kernel, with the options -j 2 for parallelism and without any -j option
for the sequential case.  In both cases all of the files should be
considered an a message should be printed at the end about the failure.

This test requires a subdirectory with a good number of .c files,
ideally some of which contain calls to kmalloc.