adjust res file for changed attribute semantics
A type metavaraible T poses a problem for attribute placement, because
attribute placement has a special behavior around the base type, but
T can match both the base type and the base type followed by any number
of *s.  So the equivalence of just after the base type and just before
the base type is not triggered in this case.  Thus one of the cases is
not matched.