Add PreEnd, to allow adding things after a function definition.
The problem with function definitions is that the final } is shared among
all returns.  So it is not suitable as an attachment point for code that is
to be added after the function.

Exit itself is also not suitable due to the self loop on Exit.  This break
check_reachability.

Add a FakeInfo on which to attach things, and put it in a new PreExit node
that does not have a loop.  The Exit node comes afterwards to fulfill the
infinite path requirement of CTL.

A new constraint CstrExit is added to match PreExit nodes.