adjust treatment of newlines in adjust_after_brace
adjust_after_brace removes extra newlines that arive when removing the
first or last line of code, if that line is preceeded or followed by
more newlines.

Some attention is needed for the newlines at the extremities, because we
always want to keep the last newline, which is followe dby the right
number of spaces for whatever comes after it.

When checking the code following a {, we want to drop the first newline
and keep the last one.

When checking the code preceeding a }, the tokens are all reversed, so
we want to drop what is in the reversed situation the last newline and
keep the first one.

improvement for csw.cocci