simplify and specialize treatment after and before brace
After an open brace, we want to remove any newlines after any initially removed code.

Before a closed brace, we want to remove any newlines after any finally remoed code.

Putting these into one function is awkward, so create a special case for each one.

Also remove all whitespace when some code is completely removed.

A special case is needed for a token that is just "".  This can occur at
the end of a #define, and should be removed when the #defined is
completely removed.