Limit trace-cmd pipeline concurrency and use artifact for clean
`resource_group: trace-cmd-vm` ensures that there is only one
trace-cmd pipeline in parallel, not to exceed CI quotas for VMs.

We use an artifact to store in VM creation job the datas needed
for the clean job, so that the latter can run without cloning the
repository, in order to work even if the branch has been deleted.