avoid matching NoType
We use NoType in order to be able to see macro parameters as local
variables.  But NoType should not cause the expression to be treated
as having a type.  So don't cause a binding to NoType in cocci_vs_c
and don't propagate NoType to the types of more complex expressions,
eg &var, where var is the name of a macro parameter.