Fix libdir in globals/config.ml.in
When rewriting the `$exec_prefix` part of libdir, use the actual value
of exec_prefix, rather than the parent directory of bindir. Although
bindir is ${exec_prefix}/bin by default, it can be overridden by passing
an argument to ./configure, in which case the paths to standard.iso and
standard.h would be deduced incorrectly.

Fixes: 540888ff426e ("Fix 263: wrong default path for COCCINELLE_HOME")

Signed-off-by: Maxim Mikityanskiy <maxtram95@gmail.com>