make a proper stack for collecting variable information for type inference
The original type inference checked for shadowing of variables by
putting all of the variables in the topmost stack frame.  This poses a
problem for C++ strcut variants.  While for C structs, the fields are
labelled as DeclarationField, for C++ structs they are labelled as
DeclFied.  A DeclField just contains a decl, because C++ fields can have
all of the features of top level declarations, such as initial values.
But then if one has a global variable and a field with the same name,
the field will seem to shadow the information about the global variable,
because the same kdecl visitor function is used (which is not used in
the case of a DeclarationField).  Here we avoid the problem, by making a
proper stak in the case of variables and searching through the while
stack, not just the top, for variables.

This loses the ability to complain about shadowing in nested scopes.
That ability might be useful due to the fact that for inherited
metsvariables, Coccinelle doesn't know where their contained variables
are declared, so entering a new scope for the same variable could be a
concern.

The test cases have to be tested with --type-c.