take scripts on metavariables into account
Need to search through the whole semantic patch to find OCaml code, if
OCaml scripting is not supported.  Add a function for traversing the
semantic patch code to look for OCaml scripts in constraints.  At the
same time, chane the searching function to return a boolean rather than
accumulating the scripts, since the list of accumulated scripts is not
used.