Better handling of ... in #define
This commit adds a special PreExit before EndNode to match ...
in #define. This is to match ... in #define. This is because ...
requires After or PreExit/ErrorExit.

c.f.
translation of `#define f(a) ...` in SmPL to CTL
```
((Ex _v . #define f(a)) &,
 AX(A[((Ex _v . _S5) & !After) U
    ((PreExit v ErrorExit) v (After v ErrorExit))]))
```

translation of `#define f(a) g(a) in SmPL to CTL
```
((Ex _v . #define f(a)) &, AX((Ex _v . g(a))))