Allow metavariables for template types and template instantiations
For a template type, the name of a template is another type, to allow the
use of metavariables.  However, in practice, this name will always either
be a type name (or a type metavariable in the semantic patch world).

The conversion to detect TemplateStart and TemplateEnd, in parsing_hacks
and in parse_cocci, takes care of ensuring this invariant.