fix parsing of C++ fields
C++ fields are parsed like ordinary variable declarations, for more
flexibility (storage, initial values etc).  But to be transformed, they
need to be converted back in to the standard representation of a
structure field.  This commit makes that change.

Note that the C++ parser accepts a lot more for structure fields than
the SmPL parser.  The SmPL parser currently only accepts the C syntax
for structure fields, so no storage or initial values.  Fields with this
extra information can be matched with field metavariables, but not with
actual patterns.  That is, there is no isomorphism that ignores storage
or initialization for fields.