Drop type information for inherited metavariables
If a previous rule changes types, the type will be out of date.  There
is no obvious way to check for this, and in particular to react in a
good way when the types do change, so just drop the type information.

The need for this is illustrated by the autoloop test case proposed by
Michele Martone <michele.martone@posteo.de>