allow removal of } after break and continue
Previously, the code for break and continue took care of checking the loop lael
but then ignored th continuation.  Handle break and continue like other atomic things,
but integrate the check for the loop label.

In the test case, the old code would leave extra braces, resulting in
code outside of any function.