extend quantifiers over dots
The encoding of A ... B will put !A and !B in place of the ... unless
when any is used.  Since A is at the root of the path, its bindings
should be shared with the bindings in !A.  On the other hand, we
continue to leave the metavariables of B out of the metavariables
quantified in the path, to allow B's metavariables to have different
values at different reachable locations.

Furthermore, statement metavariables in A and B should be taken into
account in the shortest path checks.

The latter issue was detected by an example provided by Markus Elfring <Markus.Elfring@web.de>