fix type deduction of decimal constants
In a previous commit 'is_long_dec' changed to lexicographical
comparison of strings, which will give incorrect results. It only
happens if either --int-bits or --long-bits option is given, otherwise
everything is treated as "int".

May still not be entirely correct, the standard says unsuffixed
decimal constants are one of "int", "long", or "long long", and now it
still assigns unsigned variants.

Signed-off-by: Akos Pasztory <akos.pasztory@gmail.com>