hide output of cd
CDPATH causes cd to sometimes print the name of the destination direction
in the case of cd.  Redirect the output to /dev/null to avoid confusion
with the subsequently generated list of file names.

The alternate solution of removing CDPATH from the environment was not used
because OCaml doesn't seem to have good support for this.

Reported-by: Wolfram Sang <wsa+renesas@sang-engineering.com>