Install AppStream metadata file
https://en.wikipedia.org/wiki/AppStream

AppStream a standard metadata format for describing a software project.
It contains items such as textual description, categories, authors, links
and information about releases. It is mainly used by Linux package managers.
Notably, it is required to submit the project to Flathub repository.

It is preferred to ship the AppStream upstream so that other distributions can use it:
https://docs.flathub.org/docs/for-app-authors/requirements#acceptable-but-should-be-submitted-upstream

Signed-off-by: Jan Tojnar <jtojnar@gmail.com>