Remove pos variables from the extended fv list for ...
If we have

foo@p();
...
bar();

the when code on the ... should not include p, if p is a local variable.
We want to check that there is no code of the given form, not just the code
at the initial position.

Most of the change is to pass around the list of position variables
declared in the current rule.