break up envs for e1 & e2
Since 362937b2a84840e68ae021171df10c7a4cc6fbef, e1 ... e2 has the
quantifiers for the free variables of e1 around the whole thing, to ensure
that the when code on the ... refers to the same variables as e1.  This can
make the semantic patch very slow, as illustrated by kmerr.cocci in
scripts/coccinelle/null/kmerr.cocci in the Linux kernel.  The slowness
comes from environments based on multipl metavariable bindings getting very
large.

To reduce (but not solve) the problem, for the first & where the left side
has multiple results, consider these results individually when working on
the right side, and then union the results.  This may lead to some loss of
sharing.  Maybe it is not advantageous when the ... contains when any and
does not contain any explicit when clause containing the variables of e1.

Reported-by: "Ricardo Ribalda" <ribalda@chromium.org>