advance on while loops with declarations
This compiles but it causes many test cases to fail, so it is a
transitory version.

Among the changes, Ast_cocci.WhileDecl is updated to be an
annotated_decl to allow information not mentioned in the
semantic patch to also be deleted in the allminus case.

Also consider both WhileExp and WhileDecl when searching for
tests that should not appear.