properly label #pragma node in the CFG
The node needs a label like other nearby statements, so it can be seen
as having the same nesting depth as the other statements, and thus get
removed when an enclosing block is removed.

Slightly update the test case to better illustrate what the label should
be.

The need for this is illustrated by the test case
dots_in_block_remove