move #spatch .cocci files to the end
There is nothing that undoes the effect of #spatch, and this can impact the
results on files that are tested later.  Move all such files to the end to
lessen the impact.  This is not a real solution.  Perhaps testing such
files in a separate process would be a better solution.