drop collected information on #undef
Different definitions of a macro can change the relevant hints.  #undef is
an indication that the current definition of a macro is no longer useful,
so remove any hint information about it.

Note that the hint information is global while the parsing of a file is
specific to a .c file and its headers, so this may reduce the amount of
code that is parsed successfully in a random way.

With Linux commit f66bc387efbee59978e076ce9bf123ac353b389c, the problem can
be reproduced with spatch sp.cocci ~/linux/include/trace/events
--include-headers where sp.cocci is any semantic patch that doesn't include
any special identifiers or keywords.

Reported-by: Petr Tesarik
Closes: github issue #399