Support markers among structure fields.
The scheduler task_struct contains the markers
randomized_struct_fields_start and randomized_struct_fields_end, with no
following semicolon.  These are now supported by the C parser.  They are
not supported by the semantic patch parser, but the test case shows how one
can match a field metavariable to a marker and then use python to check
that the right one has been matched.  This is sufficient to eg introduce a
new field in the randomized region.

Suggested-by: devnoname120 <devnoname120@gmail.com>