avoid expanding macros from standard.h in attribute name lists
standard.h is for macros that are essential to enable parsing.  Any name is
acceptable in an attribute list, so there is no need to expand such names.
This allows eg having an attribute name that can be used in
__attribute__((...)) and that can also be used freestanding, and to cause
the freestanding version to disappear via standard.h, if it is not
supported by Coccinelle.