forbid only fresh metavars when they come from the same rule
The parameter list of a script that initializes a fresh metavariable
can refer to metavariables declared in the same rule, because thay are
bound by matching, and thus have values when fresh metavariables are
used, at transformation time.  But fresh metavariables declared in the
same rule don't have this property, so cannot be used.

This patch stores the name of the current rule in the information
collected by the SmPL parser about each script, and then compares the
name to the rule name of each parameter that refers to a fresh
metavariable.  A parameter that refers to a fresh metavariable from
the same rule is rejected.

But one could also wonder why the parser couldn't make the same check when
it checks for the validity of a metavariable reference.  To explore.  In
any case, it seems cleaner for the metavariables in the collected list of
scripts to have their proper rule names.

Reported-by: Michele Martone