Choose a better tabbing unit
This fixes a problem when the previous line is indented by eg three
units and the current line is indented by only one (ie a difference of
more than one).  In this case, the following is now done to obtain the
tabbing unit.

1. First, there is a search from the indentation of the current line
(depthmin) plus one to the indentation of the previous line (dmin) to
find the extra indentation I associated with the smallest one available.

2. Then, there is a check whether N copies of the old tabbing unit, if
any, is the same as I, where N is the difference between the chosen
number and depthmin.  If so, the old tabbing unit is reunsed.

3. If not, I is divided by the difference between the chosen number and
depthmin, and that many characters are taked from the front of I.

Suggested-by: Erick Karanja <karanja99erick@gmail.com>